update receipt generation

+++ app/src/main/java/com/jsoft/pos/ui/utils/ImageUtil.kt
@@ -26,13 +26,15 @@ object ImageUtil {
 
             val dir = context.getDir("item_image", Context.MODE_PRIVATE)
 
-            name?.takeIf { it.isNotEmpty() }.apply {
-                val file = File(dir, name)
-                val deleted = file.delete()
-            }
-
             if (!dir.exists()) {
                 dir.mkdirs()
+            }
+
+            name?.also {
+                val file = File(dir, it)
+                if (file.exists()) {
+                    val deleted = file.delete()
+                }
             }
 
             val format = SimpleDateFormat("yyyyMMddhhmmss", Locale.ENGLISH)
@@ -56,9 +58,9 @@ object ImageUtil {
     }
 
     fun deleteImage(context: Context, name: String?) {
-        name?.takeUnless { it.isEmpty() }?.also {
+        name?.also {
             val dir = context.getDir("item_image", Context.MODE_PRIVATE)
-            val file = File(dir, name)
+            val file = File(dir, it)
 
             if (file.exists()) {
                 file.delete()
@@ -66,7 +68,7 @@ object ImageUtil {
         }
     }
 
-    fun generateReceipt(context: Context, bitmap: Bitmap): Uri? {
+    fun generateTempReceipt(context: Context, bitmap: Bitmap?): Uri? {
 
         var fos: FileOutputStream? = null
 
@@ -76,22 +78,78 @@ object ImageUtil {
 
             val format = SimpleDateFormat("yyyyMMddhhmmss", Locale.ENGLISH)
             val file = File(dir, String.format(Locale.ENGLISH, "%s.png", format.format(Date())))
+            file.deleteOnExit()
 
             fos = FileOutputStream(file)
 
-            bitmap.compress(Bitmap.CompressFormat.PNG, 100, fos)
+            bitmap?.compress(Bitmap.CompressFormat.PNG, 100, fos)
 
             return Uri.fromFile(file)
 
         } catch (e: Exception) {
             e.printStackTrace()
-            AlertUtil.showToast(context, "Error writing image")
+            AlertUtil.showToast(context, "Error generating receipt")
         } finally {
             fos?.close()
         }
 
         return null
 
+    }
+
+    fun generateReceipt(context: Context, bitmap: Bitmap, name: String?): String? {
+
+        var fos: FileOutputStream? = null
+
+        try {
+
+            val dir = context.getExternalFilesDir("receipts")
+
+            if (!dir.exists()) {
+                dir.mkdirs()
+            }
+
+            name?.also {
+                val file = File(dir, it)
+                if (file.exists()) {
+                    val deleted = file.delete()
+                }
+            }
+
+            val format = SimpleDateFormat("yyyyMMddhhmmss", Locale.ENGLISH)
+            val file = File(dir, String.format(Locale.ENGLISH, "%s.png", format.format(Date())))
+
+            fos = FileOutputStream(file)
+
+            bitmap.compress(Bitmap.CompressFormat.PNG, 100, fos)
+
+            return file.name
+
+        } catch (e: Exception) {
+            e.printStackTrace()
+            AlertUtil.showToast(context, "Error generating receipt")
+        } finally {
+            fos?.close()
+        }
+
+        return null
+
+    }
+
+    fun readReceipt(context: Context, name: String?): Uri? {
+        if (!name.isNullOrEmpty()) {
+
+            val dir = context.getExternalFilesDir("receipts")
+            val file = File(dir, name)
+
+            if (file.exists()) {
+                return Uri.fromFile(file)
+            }
+        }
+
+        AlertUtil.showToast(context, "Error loading receipt")
+
+        return null
     }
 
     fun readImage(context: Context, name: String?): Bitmap? {
